r9 - monitor adding function for issue view page
+ Requires changes made in 3 files.
+ Prevent duplication added so the form does ignore users that already
are monitoring the issue.

Tune-up:
- username can be replaced to users' actual, real name.
- Customizing the input field can be made using CSS for
'demo-input-facebook-theme' or for another input id of your creation in
order to seperate the view of this input field and the one existing on
the issue creating page.
- To improve the look of the ADD button, default.css is where changes
can be made for that.

- TODO: the monitor adding section currently does not look native with 2
fields of monitors names: the existing and the newcomers (feeder tray
with ADD button).

Workaround to this issue includes:
Doing away with 'add' button & pre-populating the existing in the Ajax-
Javascript form (b'cuz only this javascript field can display changes
made to the section dynamically).