r10+ user Real Name mesh 10.1
fix the following:
- access level can now accept the $g_default_new_account_access_level
pre-defined in config_inc (see default cfg file).. so users are created
with the access level (user class) that we deem they should belong to.
This (the setting) applies to all newly created users! beware. By
default, new user receives reporter access.
- whether they should be protected user.
- whether they should be enabled right from registration without email
activation (set it to true as it was if hardcode is preferred). With
this fix, only after use they their email to activate the account is
then set enabled and ready for use/logon.
Note: Partial testing was conducted. However, a full testing for every
of these deems necessary.