r12 categoryId
- change category_id to 1 in gpc_get_int( 'category_id', 1 ); applying
the change in line 120 to set first project chosen and save users from
having to select category each time they report new issue, knowing that
the first project is always selected. If developers or other specific
group of users want to report for new type of issue, they would just
have to open the scroll and choose issue category that is other than the
first.
- clean up all trailing spaces in the code.
- sync code of this bug_report_page.